sistema-recomendacao/requirements.txt
streamlit
pandas
numpy
scikit-surprise
plotly
